Framework mostly finished
Most of the framework is complete and is now being tested.

Several bugs will need to be addressed including error handling to
override PHP's default error handler, issues with authentication,
and behavior upon certain failures such as a view not found or an
object that "does not exist" when being accessed via a method call.